introduce classes for each command to avoid long conditional statements
Up until now we check the option passed to command and execute the
appropriate function using a lengthly conditional statement.

We change this by making a class for each command and have methods
to run for each passed option.
Since all options are mutually exclusive it is easy to map them to
a class method and call the method directly.